changed to BigInteger

+++ src/edu/vt/NetInf/HitInfo.java
@@ -1,11 +1,13 @@
 package edu.vt.NetInf;
 
+import java.math.BigInteger;
+
 public class HitInfo {
     public String NId,Parent;
-    public Long unixTime;
+    public BigInteger unixTime;
 
 
-    public HitInfo(String NId, Long unixTime) {
+    public HitInfo(String NId, BigInteger unixTime) {
         this.NId = NId;
         Parent = null;
         this.unixTime = unixTime;
